+++ misc/lfsr/xorrot32_find.py
@@ -1,8 +1,9 @@
 import lfsr_engine as lfsr
-import sympy
+import sympy, galois
+import numpy as np
 
 gen = lfsr.XorGenMaker(32)
-abc = range(1,32)#list(filter(lambda x: sympy.isprime(x), range(3, 32)))
+abc = range(1,32)
 print(abc)
 for a in abc:
     for b in abc:
@@ -10,5 +11,9 @@ for a in abc:
         for c in abc:
             if c > b:
                 T = gen.make_xorrot_matrix(a, b, c)
-                if lfsr.is_full_period(T, False):
-                    print("=====>", a, b, c)
+                Tout = np.linalg.matrix_power(galois.GF(2)(T), 2**32)
+                is_candidate = (Tout == T).all()
+                if is_candidate:
+                    print("????>", a, b, c)
+                    if lfsr.is_full_period(T, False):
+                        print("=====>", a, b, c)